+++ Examples/cloud_dedicated_query.py
@@ -1,10 +1,13 @@
+from config import Config
 import influxdb_client_3 as InfluxDBClient3
 
+config = Config()
+
 client = InfluxDBClient3.InfluxDBClient3(
-    token="",
-    host="b0c7cce5-8dbc-428e-98c6-7f996fb96467.a.influxdb.io",
-    org="6a841c0c08328fb1",
-    database="flight2")
+    token=config.token,
+    host=config.host,
+    org=config.org,
+    database=config.database)
 
 table = client.query(
     query="SELECT * FROM flight WHERE time > now() - 4h",