+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.9.0 [unreleased]
 
+### Features
+
+1. [#108](https://github.com/InfluxCommunity/influxdb3-python/pull/108): Better expose access to response headers in `InfluxDBError`.  Example `handle_http_error` added.
+
 ## 0.8.0 [2024-08-12]
 
 ### Features